Refactor: move icons to src/icons
