+++ packages/slice-machine/src/icons/LogoIcon.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+
+const LogoIcon: React.FC<React.SVGProps<SVGSVGElement>> = (props) => {
+  return (
+    <svg
+      width="32"
+      height="32"
+      viewBox="0 0 32 32"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <rect width="32" height="32" rx="6" fill="#1A1523" />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M16.2519 22.6786C19.7531 22.6786 22.5913 19.8404 22.5913 16.3393C22.5913 14.8061 22.0469 13.3999 21.141 12.3037L23.8238 10.7547C23.9923 10.6574 24.05 10.4421 23.9528 10.2736C23.8555 10.1052 23.6401 10.0474 23.4717 10.1447L20.65 11.7738C19.5101 10.6754 17.9599 10 16.2519 10C12.7508 10 9.91263 12.8382 9.91263 16.3393C9.91263 17.7073 10.3459 18.9741 11.0828 20.0099L8.17615 21.688C8.0077 21.7853 7.94999 22.0006 8.04724 22.1691C8.1445 22.3375 8.35989 22.3953 8.52834 22.298L11.5276 20.5664C12.6883 21.8628 14.3749 22.6786 16.2519 22.6786ZM11.5276 20.5664C11.3692 20.3895 11.2207 20.2038 11.0828 20.0099L17.3262 16.4052C17.4946 16.308 17.71 16.3657 17.8073 16.5341C17.9045 16.7026 17.8468 16.918 17.6784 17.0152L11.5276 20.5664ZM20.65 11.7738L14.3216 15.4275C14.1532 15.5247 14.0955 15.7401 14.1927 15.9086C14.29 16.077 14.5054 16.1347 14.6738 16.0375L21.141 12.3037C20.9874 12.1178 20.8235 11.9409 20.65 11.7738Z"
+        fill="white"
+      />
+    </svg>
+  );
+};
+
+export default LogoIcon;
